+++ integration/spark/src/main/common/java/io/openlineage/spark/agent/lifecycle/BaseVisitorFactory.java
@@ -17,7 +17,6 @@ import io.openlineage.spark.agent.lifecycle.plan.QueryPlanVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.SaveIntoDataSourceCommandVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.wrapper.InputDatasetVisitor;
 import io.openlineage.spark.agent.lifecycle.plan.wrapper.OutputDatasetVisitor;
-import io.openlineage.spark.agent.lifecycle.plan.wrapper.OutputDatasetWithMetadataVisitor;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -59,19 +58,16 @@ abstract class BaseVisitorFactory implements VisitorFactory {
         getCommonVisitors(sqlContext, jobNamespace);
     List<QueryPlanVisitor<LogicalPlan, OpenLineage.OutputDataset>> list = new ArrayList<>();
 
-    list.add(new OutputDatasetWithMetadataVisitor(new InsertIntoDataSourceDirVisitor()));
+    list.add(new OutputDatasetVisitor(new InsertIntoDataSourceDirVisitor()));
+    list.add(new OutputDatasetVisitor(new InsertIntoDataSourceVisitor(allCommonVisitors)));
+    list.add(new OutputDatasetVisitor(new InsertIntoHadoopFsRelationVisitor()));
     list.add(
-        new OutputDatasetWithMetadataVisitor(new InsertIntoDataSourceVisitor(allCommonVisitors)));
-    list.add(new OutputDatasetWithMetadataVisitor(new InsertIntoHadoopFsRelationVisitor()));
-    list.add(
-        new OutputDatasetWithMetadataVisitor(
+        new OutputDatasetVisitor(
             new SaveIntoDataSourceCommandVisitor(sqlContext, allCommonVisitors)));
     list.add(new OutputDatasetVisitor(new AppendDataVisitor(allCommonVisitors)));
     list.add(new OutputDatasetVisitor(new InsertIntoDirVisitor(sqlContext)));
     if (InsertIntoHiveTableVisitor.hasHiveClasses()) {
-      list.add(
-          new OutputDatasetWithMetadataVisitor(
-              new InsertIntoHiveTableVisitor(sqlContext.sparkContext())));
+      list.add(new OutputDatasetVisitor(new InsertIntoHiveTableVisitor(sqlContext.sparkContext())));
       list.add(new OutputDatasetVisitor(new InsertIntoHiveDirVisitor()));
     }
     return list;