+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/OpenLineage/OpenLineage/compare/0.3.1...HEAD)
 
+### Added
+* Spark output metrics [@OleksandrDvornik](https://github.com/OleksandrDvornik)
+
+### Fixed
 * dbt: column descriptions are properly filled from metadata.json [@mobuchowski](https://github.com/mobuchowski)
 * dbt: allow parsing artifacts with version higher than officially supported  [@mobuchowski](https://github.com/mobuchowski)
 * dbt: dbt build command is supported  [@mobuchowski](https://github.com/mobuchowski)